+++ class-mb-beaver-themer-integrator.php
@@ -40,10 +40,16 @@ class MB_Beaver_Themer_Integrator {
 				'type'    => 'select',
 				'label'   => __( 'Field Name', 'meta-box-beaver-themer-integrator' ),
 				'options' => $this->get_post_fields(),
+				'toggle'  => $this->get_toggle_rules(),
 			),
 			'image_size' => array(
 				'type'  => 'photo-sizes',
 				'label' => __( 'Image Size', 'meta-box-beaver-themer-integrator' ),
+			),
+			'date_format' => array(
+				'type'        => 'text',
+				'label'       => __( 'Date Format', 'meta-box-beaver-themer-integrator' ),
+				'description' => __( 'Enter a <a href="http://php.net/date">PHP date format string</a>. Leave empty to use the default field format.', 'meta-box-beaver-themer-integrator' ),
 			),
 		) );
 	}
@@ -64,6 +70,7 @@ class MB_Beaver_Themer_Integrator {
 		switch ( $field['type'] ) {
 			case 'image':
 			case 'image_advanced':
+			case 'image_upload':
 			case 'plupload_image':
 				$value = rwmb_get_value( $field_id );
 				return array_keys( $value );
@@ -72,6 +79,12 @@ class MB_Beaver_Themer_Integrator {
 				$value        = rwmb_get_value( $field_id, $args );
 				$value['id']  = $value['ID'];
 				return $value;
+			case 'date':
+			case 'datetime':
+				if ( ! empty( $settings->date_format ) ) {
+					$args['format'] = $settings->date_format;
+				}
+				break;
 		}
 
 		$value = rwmb_the_value( $field_id, $args, '', false );
@@ -86,17 +99,11 @@ class MB_Beaver_Themer_Integrator {
 	 */
 	public function get_post_fields() {
 		$sources = array();
-		$fields  = rwmb_get_registry( 'field' )->get_by_object_type( 'post' );
+		$fields  = $this->get_all_fields();;
 		foreach ( $fields as $post_type => $list ) {
 			$post_type_object = get_post_type_object( $post_type );
-			if ( ! $post_type_object ) {
-				continue;
-			}
 			$options = array();
 			foreach ( $list as $field ) {
-				if ( in_array( $field['type'], array( 'heading', 'divider', 'custom_html', 'button' ), true ) ) {
-					continue;
-				}
 				$options[ $field['id'] ] = $field['name'] ? $field['name'] : $field['id'];
 			}
 			$sources[ $post_type ] = array(
@@ -106,5 +113,64 @@ class MB_Beaver_Themer_Integrator {
 		}
 
 		return $sources;
+	}
+
+	/**
+	 * Get toggle rules for select field.
+	 * Only show additional fields when field type matches.
+	 *
+	 * @return array
+	 */
+	public function get_toggle_rules() {
+		$fields  = $this->get_all_fields();
+		$field_map = array();
+		foreach ( $fields as $post_type => $list ) {
+			foreach ( $list as $field ) {
+				$field_map[ $field['id'] ] = $field['type'];
+			}
+		}
+
+		$rules       = array();
+		$image_rules = array( 'fields' => array( 'image_size' ) );
+		$date_rules  = array( 'fields' => array( 'date_format' ) );
+		foreach ( $field_map as $id => $type ) {
+			switch ( $type ) {
+				case 'image':
+				case 'image_advanced':
+				case 'image_upload':
+				case 'plupload_image':
+				case 'single_image':
+					$rules[$id] = $image_rules;
+					break;
+				case 'date':
+				case 'datetime':
+					$rules[$id] = $date_rules;
+					break;
+			}
+		}
+
+		return $rules;
+	}
+
+	/**
+	 * Get all fields that have values.
+	 *
+	 * @return array
+	 */
+	protected function get_all_fields() {
+		$fields = rwmb_get_registry( 'field' )->get_by_object_type( 'post' );
+
+		// Remove fields for non-existing post types.
+		$fields = array_filter( $fields, function( $post_type ) {
+		    return post_type_exists( $post_type );
+		}, ARRAY_FILTER_USE_KEY );
+
+		// Remove fields that don't have value.
+		array_walk( $fields, function ( &$list ) {
+			$list = array_filter( $list, function( $field ) {
+				return ! in_array( $field['type'], array( 'heading', 'divider', 'custom_html', 'button' ), true );
+			} );
+		} );
+		return $fields;
 	}
 }