+++ CHANGELOG.md
@@ -5,5 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## Unreleased
 
-### 1.0.0
+## 1.1.0
+### Added
+- Added date format for `date`, `datetime` fields
+- Added field dependency, showing extra params only for correct field types
+
+## 1.0.0
 - First release